backend/requirements.txt
flask
flask-cors
pandas
requests
langdetect
youtube-transcript-api
google-api-python-client
python-dotenv